fix: move profiler block into ContainerDetailScreen scope + fix ramPercent type mismatch
- Moved profiler button block from DxvkConfigDialog (no containerId access)
  into ContainerDetailScreen GENERAL tab where containerId is in scope
- Fixed ProfilerSession ramPercent: cast HudMetrics.ramPercent Float to Int
- Removed duplicate TextStyle/FontWeight imports

+++ app/src/main/java/com/winlator/star/profiler/ProfilerSession.kt
@@ -106,7 +106,7 @@ class ProfilerSession(
                         cpuPercent = cpu,
                         gpuLoadPercent = gpuLoad,
                         vramUsedBytes = vram,
-                        ramPercent = ram
+                        ramPercent = ram.toInt()
                     )
                 )
